Added html elements specific module
